+++ CotEditor/Sources/SyntaxStyle.swift
@@ -331,6 +331,9 @@ extension SyntaxStyle {
         //      (2016-11, macOS 10.12.1 SDK)
         let string = textStorage.string.immutable
         
+        // avoid parsing twice for the same string
+        guard (self.syntaxHighlightParseOperationQueue.operations.last as? SyntaxHighlightParseOperation)?.string != string else { return }
+        
         self.highlight(string: string, range: wholeRange, completionHandler: completionHandler)
     }